fix(CommonInclude): countbits(unsigned long) returns unsigned long on Linux
Changed return type from unsigned long long to
unsigned long to match input type and match
Windows version consistency.

+++ WickedEngine/CommonInclude.h
@@ -287,7 +287,7 @@ inline unsigned int countbits(unsigned int value)
 {
 	return __builtin_popcount(value);
 }
-inline unsigned long long countbits(unsigned long value)
+inline unsigned long countbits(unsigned long value)
 {
 	return __builtin_popcountl(value);
 }
